requirements.txt
Flask==1.1.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==1.0.1
regex==2020.11.13
allennlp==1.3.0 
boto3==1.16.40 
botocore==1.19.40 
jmespath==0.10.0 
jsonpickle==1.4.2 
overrides==3.1.0 
s3transfer==0.3.3 
sentencepiece==0.1.94 
tensorboardX==2.1 
torch==1.7.1 
typing-extensions==3.7.4.3
allennlp==1.3.0
allennlp-models==1.3.0 
conllu==4.2.1 
ftfy==5.8 
py-rouge==1.1 
word2number==1.1 
en-core-web-sm==2.3.1
transformers==4.0.1
tokenizers==0.9.4
torch==1.7.1
textblob==0.15.3
scikit-learn==0.23.2
requests==2.24.0
requests-oauthlib==1.3.0
pandas==1.1.4
numpy==1.18.5
nltk==3.5
bert-extractive-summarizer==0.5.2